+++ projects/hex/CMakeLists.txt
@@ -6,6 +6,7 @@ enable_testing()
 set(HEX_SOURCES
         src/hex/main.cpp
         src/hex/core/cli.cpp
+        src/hex/core/disassembly.cpp
 
         src/hex/vm/slice.cpp
         src/hex/vm/virtual-machine.cpp
