+++ projects/hex/src/hex/main.cpp
@@ -1,4 +1,5 @@
 #include <hex/core/cli.hpp>
+#include <hex/core/disassembly.hpp>
 #include <hex/core/logger.hpp>
 #include <hex/vm/slice.hpp>
 #include <hex/vm/virtual-machine.hpp>
@@ -7,54 +8,12 @@
 
 #include <magic_enum/magic_enum.hpp>
 
-#include <vector>
-
-namespace hex {
-using namespace mana::literals;
-using namespace mana::vm;
-
-void EmitConstant(i64 offset, Value constant) {
-    Log("{:04} | {} | {}", offset, magic_enum::enum_name(Op::Constant), constant);
-}
-
-void EmitSimple(i64 offset, const Op op) {
-    Log("{:04} | {}", offset, magic_enum::enum_name(op));
-}
-
-void PrintBytecode(const Slice& c) {
-    const auto& code = c.Code();
-
-    for (i64 i = 0; i < code.size(); ++i) {
-        switch (const auto op = static_cast<Op>(code[i])) {
-            using enum Op;
-        case Constant:
-            EmitConstant(i, c.ConstantAt(code[i + 1]));
-            ++i;
-            break;
-        case Negate:
-        case Add:
-        case Sub:
-        case Div:
-        case Mul:
-        case Return:
-            EmitSimple(i, op);
-            break;
-        default:
-            Log("???");
-            break;
-        }
-    }
-}
-
-}  // namespace hex
-
 int main(const int argc, char** argv) {
-    using namespace mana::literals;
     using namespace mana::vm;
     hex::Slice slice;
 
-    const u8 a = slice.AddConstant(1.2);
-    const u8 b = slice.AddConstant(2.4);
+    const auto a = slice.AddConstant(1.2);
+    const auto b = slice.AddConstant(2.4);
 
     slice.Write(Op::Constant, a);
     slice.Write(Op::Constant, b);
