+++ projects/sigil/CMakeLists.txt
@@ -6,7 +6,7 @@ enable_testing()
 set(SIGIL_SOURCES
         src/sigil/ast/lexer.cpp
         src/sigil/ast/parser.cpp
-        src/sigil/ast/parse_tree.cpp
+        src/sigil/ast/parse-tree.cpp
 )
 
 set(SIGIL_INCLUDES
