+++ projects/sigil/include/sigil/core/logger.hpp
@@ -3,8 +3,8 @@
 #include <spdlog/sinks/stdout_color_sinks.h>
 #include <spdlog/spdlog.h>
 
-#include <mana/exit_codes.hpp>
-#include <mana/type_aliases.hpp>
+#include <mana/exit-codes.hpp>
+#include <mana/literals.hpp>
 
 #include <memory>
